Move Govspeak error class into own file
